+++ libraries/exoplayer_rtsp/src/test/java/androidx/media3/exoplayer/rtsp/reader/RtpH265ReaderTest.java
@@ -43,12 +43,11 @@ import org.junit.runner.RunWith;
 public class RtpH265ReaderTest {
 
   private static final long MEDIA_CLOCK_FREQUENCY = 90_000;
+  private static final int PACKET_SEQUENCE_NUMBER = 12345;
   private static final long AP_PACKET_RTP_TIMESTAMP = 9_000_000;
   private static final long AP_PACKET_2_RTP_TIMESTAMP = 9_000_040;
-  private static final int AP_PACKET_SEQUENCE_NUMBER = 12345;
   private static final long SINGLE_NALU_PACKET_1_RTP_TIMESTAMP = 9_000_040;
   private static final long SINGLE_NALU_PACKET_2_RTP_TIMESTAMP = 9_000_080;
-
   private static final byte[] AP_NALU_HEADER = getBytesFromHexString("6001");
   private static final byte[] NALU_1_LENGTH = getBytesFromHexString("000c");
   private static final byte[] NALU_1_INVALID_LENGTH = getBytesFromHexString("00ff");
@@ -57,11 +56,20 @@ public class RtpH265ReaderTest {
   private static final byte[] NALU_2_LENGTH = getBytesFromHexString("000e");
   private static final byte[] NALU_2_HEADER = getBytesFromHexString("4201");
   private static final byte[] NALU_2_PAYLOAD = getBytesFromHexString("1112131415161718191a1b1c");
+  private static final int FU_PACKET_SEQUENCE_NUMBER = 12342;
+  private static final long FU_PACKET_RTP_TIMESTAMP = 9_000_000;
+  private static final byte[] FU_NALU_HEADER = getBytesFromHexString("6201");
+  private static final byte[] FU_1_PACKET_1_FU_HEADER = getBytesFromHexString("a0");
+  private static final byte[] FU_1_PACKET_2_FU_HEADER = getBytesFromHexString("20");
+  private static final byte[] FU_1_PACKET_3_FU_HEADER = getBytesFromHexString("60");
+  private static final byte[] FU_1_PACKET_1_NALU_PAYLOAD = getBytesFromHexString("010203");
+  private static final byte[] FU_1_PACKET_2_NALU_PAYLOAD = getBytesFromHexString("040506");
+  private static final byte[] FU_1_PACKET_3_NALU_PAYLOAD = getBytesFromHexString("0708090a");
 
   private static final RtpPacket SINGLE_NALU_PACKET_1 =
       new RtpPacket.Builder()
           .setTimestamp(SINGLE_NALU_PACKET_1_RTP_TIMESTAMP)
-          .setSequenceNumber(AP_PACKET_SEQUENCE_NUMBER + 1)
+          .setSequenceNumber(PACKET_SEQUENCE_NUMBER + 1)
           .setMarker(true)
           .setPayloadData(Bytes.concat(NALU_1_HEADER, NALU_1_PAYLOAD))
           .build();
@@ -69,14 +77,14 @@ public class RtpH265ReaderTest {
   private static final RtpPacket SINGLE_NALU_PACKET_2 =
       new RtpPacket.Builder()
           .setTimestamp(SINGLE_NALU_PACKET_2_RTP_TIMESTAMP)
-          .setSequenceNumber(AP_PACKET_SEQUENCE_NUMBER + 2)
+          .setSequenceNumber(PACKET_SEQUENCE_NUMBER + 2)
           .setMarker(true)
           .setPayloadData(Bytes.concat(NALU_2_HEADER, NALU_2_PAYLOAD))
           .build();
 
   private static final RtpPacket VALID_AP_PACKET =
       createAggregationPacket(
-          AP_PACKET_SEQUENCE_NUMBER,
+          PACKET_SEQUENCE_NUMBER,
           AP_PACKET_RTP_TIMESTAMP,
           NALU_1_LENGTH,
           NALU_1_HEADER,
@@ -87,7 +95,7 @@ public class RtpH265ReaderTest {
 
   private static final RtpPacket VALID_AP_PACKET_2 =
       createAggregationPacket(
-          AP_PACKET_SEQUENCE_NUMBER + 1,
+          PACKET_SEQUENCE_NUMBER + 1,
           AP_PACKET_2_RTP_TIMESTAMP,
           NALU_1_LENGTH,
           NALU_1_HEADER,
@@ -98,7 +106,7 @@ public class RtpH265ReaderTest {
 
   private static final RtpPacket INVALID_AP_PACKET_EXTRA_BYTE =
       createAggregationPacket(
-          AP_PACKET_SEQUENCE_NUMBER,
+          PACKET_SEQUENCE_NUMBER,
           AP_PACKET_RTP_TIMESTAMP,
           NALU_1_LENGTH,
           NALU_1_HEADER,
@@ -110,7 +118,7 @@ public class RtpH265ReaderTest {
 
   private static final RtpPacket INVALID_AP_PACKET_MISSING_BYTE =
       createAggregationPacket(
-          AP_PACKET_SEQUENCE_NUMBER,
+          PACKET_SEQUENCE_NUMBER,
           AP_PACKET_RTP_TIMESTAMP,
           NALU_1_LENGTH,
           NALU_1_HEADER,
@@ -121,7 +129,7 @@ public class RtpH265ReaderTest {
 
   private static final RtpPacket INVALID_AP_PACKET_INVALID_NALU_LENGTH =
       createAggregationPacket(
-          AP_PACKET_SEQUENCE_NUMBER,
+          PACKET_SEQUENCE_NUMBER,
           AP_PACKET_RTP_TIMESTAMP,
           NALU_1_INVALID_LENGTH,
           NALU_1_HEADER,
@@ -131,11 +139,43 @@ public class RtpH265ReaderTest {
 
   private static final RtpPacket INVALID_AP_PACKET_SINGLE_NALU =
       createAggregationPacket(
-          AP_PACKET_SEQUENCE_NUMBER,
+          PACKET_SEQUENCE_NUMBER,
           AP_PACKET_RTP_TIMESTAMP,
           NALU_1_LENGTH,
           NALU_1_HEADER,
           NALU_1_PAYLOAD);
+
+  private static final RtpPacket VALID_FU_1_PACKET_1 =
+      createFragmentedPacket(
+          FU_PACKET_SEQUENCE_NUMBER,
+          FU_PACKET_RTP_TIMESTAMP,
+          /* marker= */ false,
+          FU_1_PACKET_1_FU_HEADER,
+          FU_1_PACKET_1_NALU_PAYLOAD);
+
+  private static final RtpPacket VALID_FU_1_PACKET_2 =
+      createFragmentedPacket(
+          FU_PACKET_SEQUENCE_NUMBER + 1,
+          FU_PACKET_RTP_TIMESTAMP,
+          /* marker= */ false,
+          FU_1_PACKET_2_FU_HEADER,
+          FU_1_PACKET_2_NALU_PAYLOAD);
+
+  private static final RtpPacket VALID_FU_1_PACKET_3 =
+      createFragmentedPacket(
+          FU_PACKET_SEQUENCE_NUMBER + 2,
+          FU_PACKET_RTP_TIMESTAMP,
+          /* marker= */ true,
+          FU_1_PACKET_3_FU_HEADER,
+          FU_1_PACKET_3_NALU_PAYLOAD);
+
+  private static final RtpPacket INVALID_FU_1_PACKET_2_INVALID_SEQUENCE =
+      createFragmentedPacket(
+          FU_PACKET_SEQUENCE_NUMBER + 3,
+          FU_PACKET_RTP_TIMESTAMP,
+          /* marker= */ false,
+          FU_1_PACKET_2_FU_HEADER,
+          FU_1_PACKET_2_NALU_PAYLOAD);
 
   private static final RtpPayloadFormat H265_FORMAT =
       new RtpPayloadFormat(
@@ -273,6 +313,100 @@ public class RtpH265ReaderTest {
         ParserException.class, () -> consume(rtpH265Reader, INVALID_AP_PACKET_SINGLE_NALU));
   }
 
+  @Test
+  public void consume_validMixedFragmentationPacketsAndSingleNALU() throws ParserException {
+    long naluPacket1PresentationTimestampUs =
+        Util.scaleLargeTimestamp(
+            (SINGLE_NALU_PACKET_1_RTP_TIMESTAMP - FU_PACKET_RTP_TIMESTAMP),
+            /* multiplier= */ C.MICROS_PER_SECOND,
+            /* divisor= */ MEDIA_CLOCK_FREQUENCY);
+    long naluPacket2PresentationTimestampUs =
+        Util.scaleLargeTimestamp(
+            (SINGLE_NALU_PACKET_2_RTP_TIMESTAMP - FU_PACKET_RTP_TIMESTAMP),
+            /* multiplier= */ C.MICROS_PER_SECOND,
+            /* divisor= */ MEDIA_CLOCK_FREQUENCY);
+    rtpH265Reader.createTracks(extractorOutput, /* trackId= */ 0);
+    rtpH265Reader.onReceivingFirstPacket(
+        VALID_FU_1_PACKET_1.timestamp, VALID_FU_1_PACKET_1.sequenceNumber);
+    consume(rtpH265Reader, VALID_FU_1_PACKET_1);
+    consume(rtpH265Reader, VALID_FU_1_PACKET_2);
+    consume(rtpH265Reader, VALID_FU_1_PACKET_3);
+    consume(rtpH265Reader, SINGLE_NALU_PACKET_1);
+    consume(rtpH265Reader, SINGLE_NALU_PACKET_2);
+
+    FakeTrackOutput trackOutput = extractorOutput.trackOutputs.get(0);
+    assertThat(trackOutput.getSampleCount()).isEqualTo(3);
+    assertThat(trackOutput.getSampleData(0))
+        .isEqualTo(Bytes.concat(NalUnitUtil.NAL_START_CODE, NALU_1_HEADER, NALU_1_PAYLOAD));
+    assertThat(trackOutput.getSampleTimeUs(0)).isEqualTo(0);
+    assertThat(trackOutput.getSampleData(1))
+        .isEqualTo(Bytes.concat(NalUnitUtil.NAL_START_CODE, NALU_1_HEADER, NALU_1_PAYLOAD));
+    assertThat(trackOutput.getSampleTimeUs(1)).isEqualTo(naluPacket1PresentationTimestampUs);
+    assertThat(trackOutput.getSampleData(2))
+        .isEqualTo(Bytes.concat(NalUnitUtil.NAL_START_CODE, NALU_2_HEADER, NALU_2_PAYLOAD));
+    assertThat(trackOutput.getSampleTimeUs(2)).isEqualTo(naluPacket2PresentationTimestampUs);
+  }
+
+  @Test
+  public void consume_mixedInvalidFragmentationPacketAndValidSingleNALU() throws ParserException {
+    long naluPacket1PresentationTimestampUs =
+        Util.scaleLargeTimestamp(
+            (SINGLE_NALU_PACKET_1_RTP_TIMESTAMP - FU_PACKET_RTP_TIMESTAMP),
+            /* multiplier= */ C.MICROS_PER_SECOND,
+            /* divisor= */ MEDIA_CLOCK_FREQUENCY);
+    long naluPacket2PresentationTimestampUs =
+        Util.scaleLargeTimestamp(
+            (SINGLE_NALU_PACKET_2_RTP_TIMESTAMP - FU_PACKET_RTP_TIMESTAMP),
+            /* multiplier= */ C.MICROS_PER_SECOND,
+            /* divisor= */ MEDIA_CLOCK_FREQUENCY);
+    rtpH265Reader.createTracks(extractorOutput, /* trackId= */ 0);
+    rtpH265Reader.onReceivingFirstPacket(
+        VALID_FU_1_PACKET_1.timestamp, VALID_FU_1_PACKET_1.sequenceNumber);
+    consume(rtpH265Reader, VALID_FU_1_PACKET_1);
+    consume(rtpH265Reader, INVALID_FU_1_PACKET_2_INVALID_SEQUENCE);
+    consume(rtpH265Reader, VALID_FU_1_PACKET_3);
+    consume(rtpH265Reader, SINGLE_NALU_PACKET_1);
+    consume(rtpH265Reader, SINGLE_NALU_PACKET_2);
+
+    FakeTrackOutput trackOutput = extractorOutput.trackOutputs.get(0);
+    assertThat(trackOutput.getSampleData(0))
+        .isEqualTo(Bytes.concat(NalUnitUtil.NAL_START_CODE, NALU_1_HEADER, NALU_1_PAYLOAD));
+    assertThat(trackOutput.getSampleTimeUs(0)).isEqualTo(naluPacket1PresentationTimestampUs);
+    assertThat(trackOutput.getSampleData(1))
+        .isEqualTo(Bytes.concat(NalUnitUtil.NAL_START_CODE, NALU_2_HEADER, NALU_2_PAYLOAD));
+    assertThat(trackOutput.getSampleTimeUs(1)).isEqualTo(naluPacket2PresentationTimestampUs);
+  }
+
+  @Test
+  public void consume_mixedWithMissingEndFragmentationPacketAndValidSingleNALU()
+      throws ParserException {
+    long naluPacket1PresentationTimestampUs =
+        Util.scaleLargeTimestamp(
+            (SINGLE_NALU_PACKET_1_RTP_TIMESTAMP - FU_PACKET_RTP_TIMESTAMP),
+            /* multiplier= */ C.MICROS_PER_SECOND,
+            /* divisor= */ MEDIA_CLOCK_FREQUENCY);
+    long naluPacket2PresentationTimestampUs =
+        Util.scaleLargeTimestamp(
+            (SINGLE_NALU_PACKET_2_RTP_TIMESTAMP - FU_PACKET_RTP_TIMESTAMP),
+            /* multiplier= */ C.MICROS_PER_SECOND,
+            /* divisor= */ MEDIA_CLOCK_FREQUENCY);
+    rtpH265Reader.createTracks(extractorOutput, /* trackId= */ 0);
+    rtpH265Reader.onReceivingFirstPacket(
+        VALID_FU_1_PACKET_1.timestamp, VALID_FU_1_PACKET_1.sequenceNumber);
+    consume(rtpH265Reader, VALID_FU_1_PACKET_1);
+    consume(rtpH265Reader, VALID_FU_1_PACKET_2);
+    consume(rtpH265Reader, SINGLE_NALU_PACKET_1);
+    consume(rtpH265Reader, SINGLE_NALU_PACKET_2);
+
+    FakeTrackOutput trackOutput = extractorOutput.trackOutputs.get(0);
+    assertThat(trackOutput.getSampleData(0))
+        .isEqualTo(Bytes.concat(NalUnitUtil.NAL_START_CODE, NALU_1_HEADER, NALU_1_PAYLOAD));
+    assertThat(trackOutput.getSampleTimeUs(0)).isEqualTo(naluPacket1PresentationTimestampUs);
+    assertThat(trackOutput.getSampleData(1))
+        .isEqualTo(Bytes.concat(NalUnitUtil.NAL_START_CODE, NALU_2_HEADER, NALU_2_PAYLOAD));
+    assertThat(trackOutput.getSampleTimeUs(1)).isEqualTo(naluPacket2PresentationTimestampUs);
+  }
+
   private static RtpPacket createAggregationPacket(
       int sequenceNumber, long timeStamp, byte[]... nalUnits) {
     return new RtpPacket.Builder()
@@ -280,6 +414,16 @@ public class RtpH265ReaderTest {
         .setSequenceNumber(sequenceNumber)
         .setMarker(true)
         .setPayloadData(Bytes.concat(AP_NALU_HEADER, Bytes.concat(nalUnits)))
+        .build();
+  }
+
+  private static RtpPacket createFragmentedPacket(
+      int sequenceNumber, long timeStamp, boolean marker, byte[]... nalUnits) {
+    return new RtpPacket.Builder()
+        .setTimestamp(timeStamp)
+        .setSequenceNumber(sequenceNumber)
+        .setMarker(marker)
+        .setPayloadData(Bytes.concat(FU_NALU_HEADER, Bytes.concat(nalUnits)))
         .build();
   }